requirements.txt
# Core web framework
fastapi==0.109.2
uvicorn==0.27.1
# Don't specify Starlette version as it's a dependency of FastAPI and will be installed automatically
pydantic==2.5.2
python-dotenv==1.0.0

# LLM providers
agno>=1.2.6
openai>=1.69.0
anthropic==0.7.7

# Content generation tools
duckduckgo-search==3.9.3
newspaper3k==0.2.8
lxml_html_clean==0.4.1
python-multipart==0.0.6
requests==2.31.0

# Utilities
tenacity==8.2.3
tiktoken==0.5.2
httpx

# API authentication
python-jose==3.3.0

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1

# Configuration
PyYAML==6.0.2
